feat: WIP - Create basic GameBoard component
Create basic GameBoard component that can build a board interface based on a nested array. Link up
simple form to dictate board size (not hooked up to anything yet. console.log's for now
