+++ src/App.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from 'react-dom';
-import { Hello, Game } from './components';
+import { GameBoard } from './components';
 import './App.css';
 
-render(<Game />, document.getElementById('root'));
+render(<GameBoard />, document.getElementById('root'));
