+++ src/components/index.js
@@ -1,1 +1,1 @@
-export { default as Game } from './Game';
+export { default as GameBoard } from './GameBoard';
